# openapi.yaml
openapi: 3.1.0
info:
  title: Pipedrive Executive Assistant API
  version: 1.0.0
servers:
  - url: https://pipedriveassis-production.up.railway.app
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - apiKeyAuth: []
paths:
...
  /activities:
    post:
      operationId: createActivity
      summary: Create activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject]
              properties:
                subject: { type: string }
                due_date: { type: string }
                due_time: { type: string }
                type: { type: string }
                deal_id: { type: integer }
                person_id: { type: integer }
                org_id: { type: integer }
                note: { type: string }
      responses:
        "200":
          description: Created activity
  /activities/{id}:
    patch:
      operationId: updateActivity
      summary: Update activity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated activity
  /activities/{id}/complete:
    post:
      operationId: completeActivity
      summary: Mark activity as completed
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Completed activity
  /notes:
    post:
      operationId: createNote
      summary: Create note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                deal_id: { type: integer }
                person_id: { type: integer }
                org_id: { type: integer }
      responses:
        "200":
          description: Created note
  /deals/{id}:
    patch:
      operationId: updateDeal
      summary: Update deal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated deal